move jinja clients to common + refactor MACRO_DEBUGGING usage in common
