+++ core/dbt/parser/macros.py
@@ -2,7 +2,8 @@ from typing import Iterable, List
 
 import jinja2
 
-from dbt.clients import jinja
+from dbt.common.clients import jinja
+from dbt.clients.jinja import get_supported_languages
 from dbt.contracts.graph.unparsed import UnparsedMacro
 from dbt.contracts.graph.nodes import Macro
 from dbt.contracts.files import FilePath, SourceFile
@@ -10,7 +11,7 @@ from dbt.exceptions import ParsingError
 from dbt.node_types import NodeType
 from dbt.parser.base import BaseParser
 from dbt.parser.search import FileBlock, filesystem_search
-from dbt.utils import MACRO_PREFIX
+from dbt.common.utils import MACRO_PREFIX
 
 
 class MacroParser(BaseParser[Macro]):
@@ -82,7 +83,7 @@ class MacroParser(BaseParser[Macro]):
             node = self.parse_macro(block, base_node, name)
             # get supported_languages for materialization macro
             if block.block_type_name == "materialization":
-                node.supported_languages = jinja.get_supported_languages(macro)
+                node.supported_languages = get_supported_languages(macro)
             yield node
 
     def parse_file(self, block: FileBlock):
