+++ core/dbt/utils.py
@@ -14,7 +14,6 @@ from pathlib import PosixPath, WindowsPath
 from dbt.common.utils import md5
 from dbt.common.events.types import RetryExternalCall, RecordRetryException
 from dbt.common.exceptions import (
-    DbtInternalError,
     RecursionError,
 )
 from dbt.exceptions import ConnectionError, DuplicateAliasError
@@ -79,38 +78,6 @@ def get_model_name_or_none(model):
     else:
         name = str(model)
     return name
-
-
-MACRO_PREFIX = "dbt_macro__"
-DOCS_PREFIX = "dbt_docs__"
-
-
-def get_dbt_macro_name(name):
-    if name is None:
-        raise DbtInternalError("Got None for a macro name!")
-    return f"{MACRO_PREFIX}{name}"
-
-
-def get_dbt_docs_name(name):
-    if name is None:
-        raise DbtInternalError("Got None for a doc name!")
-    return f"{DOCS_PREFIX}{name}"
-
-
-def get_materialization_macro_name(materialization_name, adapter_type=None, with_prefix=True):
-    if adapter_type is None:
-        adapter_type = "default"
-    name = f"materialization_{materialization_name}_{adapter_type}"
-    return get_dbt_macro_name(name) if with_prefix else name
-
-
-def get_docs_macro_name(docs_name, with_prefix=True):
-    return get_dbt_docs_name(docs_name) if with_prefix else docs_name
-
-
-def get_test_macro_name(test_name, with_prefix=True):
-    name = f"test_{test_name}"
-    return get_dbt_macro_name(name) if with_prefix else name
 
 
 def split_path(path):
